requirements.txt
albumentations==1.3.0
matplotlib==3.5.2
numpy==1.21.5
opencv_python==4.7.0.68
opencv_python_headless==4.7.0.72
pandas==1.4.4
Pillow==9.2.0
PyYAML==6.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
